+++ src/components/App.tsx
@@ -1,6 +1,7 @@
 import '@ethersproject/shims'
 
 import { ErrorBoundary, type Scope, wrap } from '@sentry/react-native'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import * as React from 'react'
 import { StyleSheet } from 'react-native'
 import { GestureHandlerRootView } from 'react-native-gesture-handler'
@@ -12,26 +13,37 @@ import { EdgeCoreManager } from './services/EdgeCoreManager'
 import { StatusBarManager } from './services/StatusBarManager'
 import { ThemeProvider } from './services/ThemeContext'
 
-function MainApp() {
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: 2,
+      refetchOnWindowFocus: false
+    }
+  }
+})
+
+const MainApp: React.FC = () => {
   const handleBeforeCapture = useHandler((scope: Scope) => {
     scope.setLevel('fatal')
     scope.setTag('handled', false)
   })
 
   return (
-    <SafeAreaProvider>
-      <ThemeProvider>
-        <GestureHandlerRootView style={StyleSheet.absoluteFill}>
-          <ErrorBoundary
-            beforeCapture={handleBeforeCapture}
-            fallback={<CrashScene />}
-          >
-            <StatusBarManager />
-            <EdgeCoreManager />
-          </ErrorBoundary>
-        </GestureHandlerRootView>
-      </ThemeProvider>
-    </SafeAreaProvider>
+    <QueryClientProvider client={queryClient}>
+      <SafeAreaProvider>
+        <ThemeProvider>
+          <GestureHandlerRootView style={StyleSheet.absoluteFill}>
+            <ErrorBoundary
+              beforeCapture={handleBeforeCapture}
+              fallback={<CrashScene />}
+            >
+              <StatusBarManager />
+              <EdgeCoreManager />
+            </ErrorBoundary>
+          </GestureHandlerRootView>
+        </ThemeProvider>
+      </SafeAreaProvider>
+    </QueryClientProvider>
   )
 }