+++ src/components/scenes/GuiPluginListScene.tsx
@@ -46,9 +46,8 @@ import { useDispatch, useSelector } from '../../types/reactRedux'
 import type { Dispatch } from '../../types/reduxTypes'
 import type { AccountReferral } from '../../types/ReferralTypes'
 import type {
-  BuyTabSceneProps,
-  NavigationBase,
-  SellTabSceneProps
+  BuySellTabSceneProps,
+  NavigationBase
 } from '../../types/routerTypes'
 import type { PluginTweak } from '../../types/TweakTypes'
 import { getPartnerIconUri } from '../../util/CdnUris'
@@ -60,6 +59,7 @@ import { bestOfPlugins } from '../../util/ReferralHelpers'
 import { logEvent, type OnLogEvent } from '../../util/tracking'
 import { base58ToUuid, getOsVersion } from '../../util/utils'
 import { EdgeCard } from '../cards/EdgeCard'
+import { PaymentOptionCard } from '../cards/PaymentOptionCard'
 import {
   EdgeAnim,
   fadeInUp20,
@@ -80,7 +80,6 @@ import {
   type ThemeProps,
   useTheme
 } from '../services/ThemeContext'
-import { DividerLine } from '../themed/DividerLine'
 import { EdgeText } from '../themed/EdgeText'
 import { SceneHeader } from '../themed/SceneHeader'
 import { SelectableRow } from '../themed/SelectableRow'
@@ -133,12 +132,12 @@ const pluginPartnerLogos: Record<string, 'guiPluginLogoMoonpay'> = {
   moonpay: 'guiPluginLogoMoonpay'
 }
 
-type BuyProps = BuyTabSceneProps<'pluginListBuy'>
-type SellProps = SellTabSceneProps<'pluginListSell'>
+type BuyProps = BuySellTabSceneProps<'pluginListBuyOld'>
+type SellProps = BuySellTabSceneProps<'pluginListSellOld'>
 type OwnProps = BuyProps | SellProps
 
 function isBuyProps(props: OwnProps): props is BuyProps {
-  return props.route.name === 'pluginListBuy'
+  return props.route.name === 'pluginListBuyOld'
 }
 
 interface StateProps {
@@ -182,7 +181,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     this.componentMounted = true
   }
 
-  async componentDidMount() {
+  componentDidMount(): void {
     this.updatePlugins()
     const { developerPluginUri } = getDeviceSettings()
     if (developerPluginUri != null) {
@@ -190,7 +189,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     }
   }
 
-  componentWillUnmount() {
+  componentWillUnmount(): void {
     this.componentMounted = false
     if (this.timeoutId != null) clearTimeout(this.timeoutId)
   }
@@ -206,7 +205,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     }
   }
 
-  updatePlugins() {
+  updatePlugins(): void {
     // Create new array objects so we aren't patching the original JSON
     const currentPlugins: BuySellPlugins = {
       buy: [...(buySellPlugins.buy ?? [])],
@@ -266,7 +265,10 @@ class GuiPluginList extends React.PureComponent<Props, State> {
   /**
    * Launch the provided plugin, including pre-flight checks.
    */
-  async openPlugin(listRow: GuiPluginRow, longPress: boolean = false) {
+  async openPlugin(
+    listRow: GuiPluginRow,
+    longPress: boolean = false
+  ): Promise<void> {
     const {
       account,
       accountReferral,
@@ -316,7 +318,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
         this.setState({ developerUri: deepPath })
 
         // Write to disk lazily:
-        writeDeveloperPluginUri(deepPath).catch(error => {
+        writeDeveloperPluginUri(deepPath).catch((error: unknown) => {
           showError(error)
         })
       }
@@ -385,7 +387,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     onPluginOpened()
   }
 
-  renderTitle = (guiPluginRow: GuiPluginRow) => {
+  renderTitle = (guiPluginRow: GuiPluginRow): React.ReactElement => {
     const styles = getStyles(this.props.theme)
     const { title, customTitleKey } = guiPluginRow
 
@@ -424,22 +426,22 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     )(error)
     if (regionError != null && regionError.length > 0) {
       const country = COUNTRY_CODES.find(c => c['alpha-2'] === countryCode)
-      const countryName = country ? country.name : countryCode // Fallback to countryCode if not found
+      const countryName = country != null ? country.name : countryCode // Fallback to countryCode if not found
 
       // Attempt to find the stateProvince name if stateProvinceCode is provided
       let stateProvinceName = stateProvinceCode
-      if (country?.stateProvinces && stateProvinceCode) {
+      if (country?.stateProvinces != null && stateProvinceCode != null) {
         const stateProvince = country.stateProvinces.find(
           sp => sp['alpha-2'] === stateProvinceCode
         )
-        stateProvinceName = stateProvince
-          ? stateProvince.name
-          : stateProvinceCode // Fallback to stateProvinceCode if not found
+        stateProvinceName =
+          stateProvince != null ? stateProvince.name : stateProvinceCode // Fallback to stateProvinceCode if not found
       }
 
-      const text = stateProvinceName
-        ? `${stateProvinceName}, ${countryName}`
-        : countryName
+      const text =
+        stateProvinceName != null
+          ? `${stateProvinceName}, ${countryName}`
+          : countryName
       Airship.show<'ok' | undefined>(bridge => (
         <ButtonsModal
           bridge={bridge}
@@ -452,7 +454,10 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     }
   }
 
-  renderPlugin = ({ item, index }: ListRenderItemInfo<GuiPluginRow>) => {
+  renderPlugin = ({
+    item,
+    index
+  }: ListRenderItemInfo<GuiPluginRow>): React.ReactElement | null => {
     const { theme } = this.props
     const { pluginId } = item
     const plugin = guiPlugins[pluginId]
@@ -462,66 +467,45 @@ class GuiPluginList extends React.PureComponent<Props, State> {
 
     const styles = getStyles(this.props.theme)
     const partnerLogoThemeKey = pluginPartnerLogos[pluginId]
-    const pluginPartnerLogo = partnerLogoThemeKey
-      ? theme[partnerLogoThemeKey]
-      : { uri: getPartnerIconUri(item.partnerIconPath ?? '') }
     const poweredBy = plugin.poweredBy ?? plugin.displayName
+    const partner =
+      poweredBy == null || item.partnerIconPath == null
+        ? undefined
+        : {
+            displayName: poweredBy,
+            icon:
+              partnerLogoThemeKey != null
+                ? theme[partnerLogoThemeKey]
+                : { uri: getPartnerIconUri(item.partnerIconPath ?? '') }
+          }
+    const [totalAmount, settlementTime] = item.description.split('\n')
     return (
       <EdgeAnim
         enter={{ type: 'fadeInDown', distance: 30 * (index + 1) }}
         style={styles.hackContainer}
       >
-        <EdgeCard
-          icon={
-            <Image
-              style={styles.logo}
-              // @ts-expect-error
-              source={theme[paymentTypeLogosById[item.paymentTypeLogoKey]]}
-            />
-          }
+        <PaymentOptionCard
+          title={this.renderTitle(item)}
+          // @ts-expect-error - we can assume paymentTypeLogoKey exists within paymentTypeLogosById because it comes from static JSON
+          icon={theme[paymentTypeLogosById[item.paymentTypeLogoKey]]}
+          totalAmount={totalAmount}
+          settlementTime={settlementTime}
+          partner={partner}
           onPress={async () => {
-            await this.openPlugin(item).catch(error => {
-              this.handleError(error)
-            })
+            await this.openPlugin(item)
           }}
           onLongPress={async () => {
-            await this.openPlugin(item, true).catch(error => {
-              this.handleError(error)
-            })
+            await this.openPlugin(item, true).catch(this.handleError)
           }}
-          paddingRem={[1, 0.5, 1, 0.5]}
-        >
-          <View style={styles.cardContentContainer}>
-            {this.renderTitle(item)}
-            {item.description === '' ? null : (
-              <EdgeText style={styles.subtitleText}>
-                {item.description}
-              </EdgeText>
-            )}
-            {poweredBy != null && item.partnerIconPath != null ? (
-              <>
-                <DividerLine marginRem={[0.25, 1, 0.25, 0]} />
-                <View style={styles.pluginRowPoweredByRow}>
-                  <EdgeText style={styles.footerText}>
-                    {lstrings.plugin_powered_by_space}
-                  </EdgeText>
-                  <Image
-                    style={styles.partnerIconImage}
-                    source={pluginPartnerLogo}
-                  />
-                  <EdgeText style={styles.footerText}>
-                    {' ' + poweredBy}
-                  </EdgeText>
-                </View>
-              </>
-            ) : null}
-          </View>
-        </EdgeCard>
+          onProviderPress={async () => {
+            await this.openPlugin(item)
+          }}
+        />
       </EdgeAnim>
     )
   }
 
-  renderTop = () => {
+  renderTop = (): React.ReactElement => {
     const {
       account,
       countryCode,
@@ -539,7 +523,9 @@ class GuiPluginList extends React.PureComponent<Props, State> {
       sp => sp['alpha-2'] === stateProvinceCode
     )
     const uri = `${FLAG_LOGO_URL}/${
-      countryData?.filename ?? countryData?.name.toLowerCase().replace(' ', '-')
+      countryData?.filename ??
+      countryData?.name.toLowerCase().replace(' ', '-') ??
+      ''
     }.png`
     const hasCountryData = countryData != null
 
@@ -619,7 +605,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     )
   }
 
-  renderEmptyList = () => {
+  renderEmptyList = (): React.ReactElement | null => {
     const { countryCode, theme } = this.props
     const styles = getStyles(theme)
     if (countryCode === '') return null
@@ -633,7 +619,7 @@ class GuiPluginList extends React.PureComponent<Props, State> {
     )
   }
 
-  render() {
+  render(): React.ReactElement {
     const {
       accountPlugins,
       accountReferral,
@@ -693,22 +679,10 @@ const getStyles = cacheStyles((theme: Theme) => ({
     // TODO: Make SceneHeader work right under UI4
     overflow: 'visible'
   },
-  cardContentContainer: {
-    flexDirection: 'column',
-    flexShrink: 1,
-    marginRight: theme.rem(0.5)
-  },
   hackContainer: {
     // HACK: Required for the header underline to span all the way to the right
     // TODO: Make SceneHeader work right under UI4
     paddingHorizontal: theme.rem(0.5)
-  },
-  selectedCountryRow: {
-    marginTop: theme.rem(1.5),
-    marginBottom: theme.rem(1.5),
-    marginHorizontal: theme.rem(1.5),
-    flexDirection: 'row',
-    alignItems: 'center'
   },
   selectedCountryFlag: {
     height: theme.rem(2),
@@ -731,18 +705,6 @@ const getStyles = cacheStyles((theme: Theme) => ({
   emptyPluginText: {
     textAlign: 'center'
   },
-  pluginRowPoweredByRow: {
-    flexDirection: 'row',
-    justifyContent: 'flex-start',
-    alignItems: 'center'
-  },
-  logo: {
-    margin: theme.rem(0.5),
-    width: theme.rem(2),
-    height: theme.rem(2),
-    aspectRatio: 1,
-    resizeMode: 'contain'
-  },
   titleText: {
     fontFamily: theme.fontFaceMedium
   },
@@ -758,20 +720,6 @@ const getStyles = cacheStyles((theme: Theme) => ({
     aspectRatio: 150 / 64,
     resizeMode: 'contain',
     marginBottom: 1
-  },
-  subtitleText: {
-    marginTop: theme.rem(0.25),
-    fontSize: theme.rem(0.75),
-    color: theme.secondaryText
-  },
-  footerText: {
-    fontSize: theme.rem(0.75),
-    color: theme.secondaryText
-  },
-  partnerIconImage: {
-    aspectRatio: 1,
-    width: theme.rem(0.75),
-    height: theme.rem(0.75)
   }
 }))
 
@@ -879,9 +827,9 @@ const GuiPluginListSceneComponent = React.memo((props: OwnProps) => {
 })
 
 // Export separate components for buy and sell routes
-export const BuyScene = (props: BuyProps) => (
+export const BuyScene = (props: BuyProps): React.ReactElement => (
   <GuiPluginListSceneComponent {...props} />
 )
-export const SellScene = (props: SellProps) => (
+export const SellScene = (props: SellProps): React.ReactElement => (
   <GuiPluginListSceneComponent {...props} />
 )