+++ src/plugins/gui/scenes/FiatPluginWebView.tsx
@@ -3,18 +3,19 @@ import * as React from 'react'
 import { WebView, type WebViewNavigation } from 'react-native-webview'
 
 import { SceneWrapper } from '../../../components/common/SceneWrapper'
+import { showError } from '../../../components/services/AirshipInstance'
 import { useHandler } from '../../../hooks/useHandler'
-import type { BuyTabSceneProps } from '../../../types/routerTypes'
+import type { BuySellTabSceneProps } from '../../../types/routerTypes'
 
 export interface FiatPluginOpenWebViewParams {
   url: string
   injectedJs?: string
   onClose?: (() => boolean) | (() => void)
   onMessage?: (message: string, injectJs: (js: string) => void) => void
-  onUrlChange?: (url: string) => void
+  onUrlChange?: (url: string) => void | Promise<void>
 }
 
-interface Props extends BuyTabSceneProps<'guiPluginWebView'> {}
+interface Props extends BuySellTabSceneProps<'guiPluginWebView'> {}
 
 export function FiatPluginWebViewComponent(props: Props): React.ReactElement {
   const { route } = props
@@ -25,10 +26,15 @@ export function FiatPluginWebViewComponent(props: Props): React.ReactElement {
 
   const handleNavigationStateChange = useHandler((event: WebViewNavigation) => {
     console.log('FiatPluginWebView navigation: ', event)
-    if (onUrlChange != null) onUrlChange(event.url)
+    if (onUrlChange != null) {
+      const p = onUrlChange(event.url)
+      p?.catch((error: unknown) => {
+        showError(error)
+      })
+    }
   })
 
-  const injectJs = (js: string) => {
+  const injectJs = (js: string): void => {
     if (webViewRef.current != null) webViewRef.current.injectJavaScript(js)
   }