+++ src/types/routerTypes.tsx
@@ -49,6 +49,8 @@ import type { MigrateWalletCalculateFeeParams } from '../components/scenes/Migra
 import type { MigrateWalletCompletionParams } from '../components/scenes/MigrateWalletCompletionScene'
 import type { MigrateWalletSelectCryptoParams } from '../components/scenes/MigrateWalletSelectCryptoScene'
 import type { OtpRepairParams } from '../components/scenes/OtpRepairScene'
+import type { RampCreateParams } from '../components/scenes/RampCreateScene'
+import type { RampSelectOptionParams } from '../components/scenes/RampSelectOptionScene'
 import type { RequestParams } from '../components/scenes/RequestScene'
 import type { SendScene2Params } from '../components/scenes/SendScene2'
 import type { EarnSceneParams } from '../components/scenes/Staking/EarnScene'
@@ -95,10 +97,18 @@ export type WalletsTabParamList = {} & {
   transactionDetails: TransactionDetailsParams
 }
 
-export type BuyTabParamList = {} & {
+export type BuySellTabParamList = {} & {
   // Buy-specific navigation
-  pluginListBuy?: GuiPluginListParams
+  pluginListBuy?: RampCreateParams
+  pluginListBuyOld?: GuiPluginListParams
   pluginViewBuy: PluginViewParams
+
+  pluginListSell?: RampCreateParams
+  pluginListSellOld?: GuiPluginListParams
+  pluginViewSell: PluginViewParams
+
+  // Ramp plugin
+  rampSelectOption: RampSelectOptionParams
 
   // Shared GUI plugin forms/displays
   guiPluginAddressForm: FiatPluginAddressFormParams
@@ -114,18 +124,14 @@ export type BuyTabParamList = {} & {
   rewardsCardWelcome: RewardsCardWelcomeParams
 }
 
-export type SellTabParamList = Omit<
-  BuyTabParamList,
-  'pluginListBuy' | 'pluginViewBuy'
-> & {
-  pluginListSell?: GuiPluginListParams
-  pluginViewSell: PluginViewParams
-}
-
 export type SwapTabParamList = {} & {
   swapCreate: SwapCreateParams | undefined
   swapConfirmation: SwapConfirmationParams
   swapProcessing: SwapProcessingParams
+}
+
+export interface TradeTabParamList {
+  pluginListBuy: RampCreateParams
 }
 
 export type EdgeTabsParamList = {} & {
@@ -133,8 +139,8 @@ export type EdgeTabsParamList = {} & {
   walletsTab:
     | NavigationCore.NavigatorScreenParams<WalletsTabParamList>
     | undefined
-  buyTab: NavigationCore.NavigatorScreenParams<BuyTabParamList> | undefined
-  sellTab: NavigationCore.NavigatorScreenParams<SellTabParamList> | undefined
+  buyTab: NavigationCore.NavigatorScreenParams<BuySellTabParamList> | undefined
+  sellTab: NavigationCore.NavigatorScreenParams<BuySellTabParamList> | undefined
   swapTab: NavigationCore.NavigatorScreenParams<SwapTabParamList> | undefined
   extraTab: undefined
   devTab: undefined
@@ -270,14 +276,9 @@ export type EdgeTabsSceneProps<Name extends keyof EdgeTabsParamList> =
     BottomTabScreenProps<EdgeTabsParamList, Name>,
     EdgeAppSceneProps<keyof EdgeAppStackParamList>
   >
-export type BuyTabSceneProps<Name extends keyof BuyTabParamList> =
+export type BuySellTabSceneProps<Name extends keyof BuySellTabParamList> =
   CompositeScreenProps<
-    StackScreenProps<BuyTabParamList, Name>,
-    EdgeTabsSceneProps<keyof EdgeTabsParamList>
-  >
-export type SellTabSceneProps<Name extends keyof SellTabParamList> =
-  CompositeScreenProps<
-    StackScreenProps<SellTabParamList, Name>,
+    StackScreenProps<BuySellTabParamList, Name>,
     EdgeTabsSceneProps<keyof EdgeTabsParamList>
   >
 export type SwapTabSceneProps<Name extends keyof SwapTabParamList> =
@@ -307,8 +308,7 @@ export type AppParamList = RootParamList &
   EdgeAppStackParamList &
   EdgeTabsParamList &
   SwapTabParamList &
-  BuyTabParamList &
-  SellTabParamList &
+  BuySellTabParamList &
   WalletsTabParamList
 
 export type RouteSceneKey = keyof AppParamList